+++ dtu/src/app/files/app/setup/AbstractTest.kt
@@ -1,4 +1,4 @@
-package {{ app_pkg }}
+package c.arve
 
 import android.content.Context
 import android.os.Bundle
@@ -10,17 +10,26 @@ abstract class AbstractTest(
     protected var logger: AbstractLogger = AndroidLogger()
 
     abstract fun doTest(extras: Bundle?): Boolean
+    open fun cleanup() { }
 
     override fun runTest(data: Bundle?, logger: ILogger?): Boolean {
         if (logger != null) {
             this.logger = WrappedLogger(logger)
         }
-        return try {
+        val success = try {
             doTest(data)
         } catch (e : Exception) {
             this.logger.error("unhandled exception", e)
             false
         }
+
+        try {
+            cleanup()
+        } catch (e : Exception) {
+            this.logger.error("cleanup failed with exception", e)
+        }
+
+        return success
     }
 
 }