+++ dtu/src/app/files/app/setup/bundleHelper.kt
@@ -1,4 +1,4 @@
-package {{ app_pkg }}
+package c.arve
 
 import android.os.Bundle
 import kotlin.reflect.KClass